Rename to test_sanitize_html
